refactor: restucture Github service functions
