+++ github_webhook_handler/src/services/github/fetchPullRequestFiles.test.ts
@@ -1,0 +1,26 @@
+import axios from 'axios'
+import { fetchPullRequestFiles } from './fetchPullRequestFiles'
+
+jest.mock("axios");
+
+jest.mock("@/config", () => ({
+  GITHUB_WEBHOOK_SECRET: "mocked-secret",
+  GITHUB_API_URL: "https://api.github.com",
+}));
+
+describe('fetchPullRequestFiles', () => {
+  it("fetches PR files successfully", async () => {
+    const mockResponse = {
+      data: [{ filename: "src/index.ts" }, { filename: "src/utils.ts" }],
+    };
+    (axios.get as jest.Mock).mockResolvedValue(mockResponse);
+
+    const files = await fetchPullRequestFiles("user/repo", 123);
+
+    expect(files).toEqual(mockResponse.data);
+    expect(axios.get).toHaveBeenCalledWith(
+      "https://api.github.com/repos/user/repo/pulls/123/files",
+      expect.any(Object)
+    );
+  });
+});
