+++ github_webhook_handler/src/services/openai/types.ts
